requirements.txt
pandas==2.2.2
streamlit==1.33.0
requests==2.23.0
streamviz==5.1
pillow==10.3.0